LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scrollable-overflow-transform-005.tentative-expected.txt


FAIL .container 1 assert_equals:
<div class="container" data-expected-scroll-width="100" data-expected-scroll-height="100">
    <div style="transform: translate(-200px, -200px);" class="element">
    </div>
  </div>
scrollHeight expected 100 but got 200
FAIL .container 2 assert_equals:
<div class="container" data-expected-scroll-width="150" data-expected-scroll-height="150">
    <div style="transform: translate(-25%, -25%);" class="element">
    </div>
  </div>
scrollHeight expected 150 but got 200
PASS .container 3
PASS .container 4
PASS .container 5
PASS .container 6
